Clean up placeholder loaders
Removed unnecessary picker usage (we always want the white constants)
Removed now-unnecessary abstractions (which I should have done when dropping 1.21.5 and lower, whoops)
